Create README.md
Object class is present in java.lang package. Every class in Java is directly or indirectly derived from the Object class. If a class does not extend any other class then it is a direct child class of Object and if extends another class then it is indirectly derived. Therefore the Object class methods are available to all Java classes. Hence Object class acts as a root of the inheritance hierarchy in any Java Program.
The Object class provides multiple methods which are as follows:
•	toString() method
•	hashCode() method
•	equals(Object obj) method
•	finalize() method
•	getClass() method
•	clone() method
•	wait(), notify() notifyAll() methods

 equals():
 If you want to compare the references of two objects, use equals() method of Object class. This method compares the references of two objects.
It returns true if the object invoking equals() method is equal to the object passed as an argument to the equals() method, otherwise returns false.

 toString(): 
This method gives a string representation of an object. For example, the string representation of an Integer object is an integer number printed as a string. 

 hashCode(): 
method returns a hash code value of an object.